Refactor debugger to have a registry
